# apps/web/openapi.json
{"components":{"schemas":{"EchoRequest":{"properties":{"message":{"default":"hello","title":"Message","type":"string"}},"title":"EchoRequest","type":"object"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"title":"Detail","type":"array"}},"title":"HTTPValidationError","type":"object"},"HealthResponse":{"properties":{"service":{"title":"Service","type":"string"},"status":{"title":"Status","type":"string"},"version":{"title":"Version","type":"string"}},"required":["status","service","version"],"title":"HealthResponse","type":"object"},"InstrumentRead":{"description":"Instrument response schema.","properties":{"created_at":{"format":"date-time","title":"Created At","type":"string"},"currency":{"title":"Currency","type":"string"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"exchange":{"title":"Exchange","type":"string"},"id":{"title":"Id","type":"string"},"industry":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Industry"},"instrument_type":{"title":"Instrument Type","type":"string"},"market":{"title":"Market","type":"string"},"name":{"title":"Name","type":"string"},"sector":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sector"},"status":{"title":"Status","type":"string"},"symbol":{"title":"Symbol","type":"string"},"updated_at":{"format":"date-time","title":"Updated At","type":"string"}},"required":["id","symbol","name","instrument_type","exchange","market","currency","status","created_at","updated_at"],"title":"InstrumentRead","type":"object"},"ResearchErrorCode":{"description":"Stable public Research API error codes.","enum":["INSTRUMENT_NOT_FOUND","RESEARCH_PACKAGE_NOT_FOUND","RESEARCH_PACKAGE_ALREADY_EXISTS","RESEARCH_VERSION_CONFLICT","IDEMPOTENCY_CONFLICT","RESEARCH_VALIDATION_ERROR","RESEARCH_PERSISTENCE_CONFLICT","RESEARCH_DOMAIN_ERROR"],"title":"ResearchErrorCode","type":"string"},"ResearchErrorDetail":{"description":"Stable public Research API error detail.","properties":{"code":{"$ref":"#/components/schemas/ResearchErrorCode"},"current_version":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Current Version"},"expected_version":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Expected Version"},"message":{"title":"Message","type":"string"}},"required":["code","message"],"title":"ResearchErrorDetail","type":"object"},"ResearchErrorResponse":{"description":"Stable public Research API error response.","properties":{"detail":{"$ref":"#/components/schemas/ResearchErrorDetail"}},"required":["detail"],"title":"ResearchErrorResponse","type":"object"},"ResearchFreshness":{"description":"Deterministic Research module freshness values.","enum":["UNVERIFIED","FRESH","STALE","FAILED"],"title":"ResearchFreshness","type":"string"},"ResearchModuleRead":{"description":"Research module snapshot response schema.","properties":{"as_of":{"format":"date-time","title":"As Of","type":"string"},"created_at":{"format":"date-time","title":"Created At","type":"string"},"freshness":{"$ref":"#/components/schemas/ResearchFreshness","description":"Deterministic content freshness. UNVERIFIED means no source-backed verification exists; FRESH/STALE are derived from last_verified_at and stale_after; FAILED means module refresh failed."},"id":{"title":"Id","type":"string"},"last_verified_at":{"anyOf":[{"format":"date-time","type":"string"},{"type":"null"}],"title":"Last Verified At"},"module_type":{"title":"Module Type","type":"string"},"module_version":{"title":"Module Version","type":"integer"},"origin_module_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Origin Module Id"},"research_package_id":{"title":"Research Package Id","type":"string"},"source_refs":{"items":{"type":"string"},"title":"Source Refs","type":"array"},"stale_after":{"format":"date-time","title":"Stale After","type":"string"},"status":{"title":"Status","type":"string"},"summary":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Summary"}},"required":["id","research_package_id","module_type","module_version","status","freshness","source_refs","as_of","stale_after","created_at"],"title":"ResearchModuleRead","type":"object"},"ResearchModuleType":{"description":"Supported Research module types.","enum":["COMPANY","BUSINESS","INDUSTRY","ORDER","FINANCIAL","EXPECTATION","VALUATION","RISK","CATALYST","COMPETITOR","MANAGEMENT"],"title":"ResearchModuleType","type":"string"},"ResearchPackageRead":{"description":"Research package version response schema.\n\nPackage status is a build lifecycle state. Module freshness is the content\nfreshness signal. PENDING or UNVERIFIED content is not ACTIVE verified\nresearch, and stale content must not be silently displayed as current.","properties":{"as_of":{"format":"date-time","title":"As Of","type":"string"},"completed_at":{"anyOf":[{"format":"date-time","type":"string"},{"type":"null"}],"title":"Completed At"},"created_at":{"format":"date-time","title":"Created At","type":"string"},"expected_version":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Expected Version"},"id":{"title":"Id","type":"string"},"instrument_id":{"title":"Instrument Id","type":"string"},"last_verified_at":{"anyOf":[{"format":"date-time","type":"string"},{"type":"null"}],"title":"Last Verified At"},"modules":{"items":{"$ref":"#/components/schemas/ResearchModuleRead"},"title":"Modules","type":"array"},"previous_version_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Previous Version Id"},"started_at":{"format":"date-time","title":"Started At","type":"string"},"status":{"description":"Build lifecycle status for the package version, not a content freshness verdict. PENDING does not mean ACTIVE verified research.","title":"Status","type":"string"},"trigger_type":{"title":"Trigger Type","type":"string"},"version":{"title":"Version","type":"integer"}},"required":["id","instrument_id","version","trigger_type","status","as_of","started_at","created_at","modules"],"title":"ResearchPackageRead","type":"object"},"ResearchRefreshRequest":{"description":"Request to create an incremental Research Package version.","properties":{"expected_version":{"minimum":1.0,"title":"Expected Version","type":"integer"},"module_types":{"items":{"$ref":"#/components/schemas/ResearchModuleType"},"maxItems":11,"minItems":1,"title":"Module Types","type":"array"}},"required":["expected_version","module_types"],"title":"ResearchRefreshRequest","type":"object"},"SystemStatusResponse":{"properties":{"api":{"additionalProperties":true,"title":"Api","type":"object"},"services":{"additionalProperties":true,"title":"Services","type":"object"},"status":{"title":"Status","type":"string"}},"required":["status","api","services"],"title":"SystemStatusResponse","type":"object"},"TaskDispatchResponse":{"properties":{"actor":{"title":"Actor","type":"string"},"message":{"title":"Message","type":"string"},"status":{"title":"Status","type":"string"},"task_id":{"title":"Task Id","type":"string"}},"required":["task_id","actor","status","message"],"title":"TaskDispatchResponse","type":"object"},"ValidationError":{"properties":{"ctx":{"title":"Context","type":"object"},"input":{"title":"Input"},"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"title":"Location","type":"array"},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}},"required":["loc","msg","type"],"title":"ValidationError","type":"object"},"WatchlistAddRequest":{"description":"Request to add an instrument to the watchlist.","properties":{"notes":{"anyOf":[{"maxLength":2000,"type":"string"},{"type":"null"}],"title":"Notes"},"query":{"maxLength":128,"minLength":1,"title":"Query","type":"string"}},"required":["query"],"title":"WatchlistAddRequest","type":"object"},"WatchlistItemRead":{"description":"Watchlist item response schema.","properties":{"agent_action":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Agent Action"},"classification":{"title":"Classification","type":"string"},"classification_confidence":{"title":"Classification Confidence","type":"integer"},"classification_reason":{"title":"Classification Reason","type":"string"},"created_at":{"format":"date-time","title":"Created At","type":"string"},"id":{"title":"Id","type":"string"},"instrument":{"$ref":"#/components/schemas/InstrumentRead"},"instrument_id":{"title":"Instrument Id","type":"string"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes"},"research_score":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Research Score"},"research_status":{"title":"Research Status","type":"string"},"thesis_summary":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Thesis Summary"},"updated_at":{"format":"date-time","title":"Updated At","type":"string"}},"required":["id","instrument_id","instrument","classification","classification_confidence","classification_reason","research_status","created_at","updated_at"],"title":"WatchlistItemRead","type":"object"},"WatchlistUpdateRequest":{"description":"Request to update a watchlist item.","properties":{"agent_action":{"anyOf":[{"maxLength":128,"type":"string"},{"type":"null"}],"title":"Agent Action"},"classification":{"anyOf":[{"maxLength":32,"type":"string"},{"type":"null"}],"title":"Classification"},"notes":{"anyOf":[{"maxLength":2000,"type":"string"},{"type":"null"}],"title":"Notes"},"research_status":{"anyOf":[{"maxLength":32,"type":"string"},{"type":"null"}],"title":"Research Status"}},"title":"WatchlistUpdateRequest","type":"object"}}},"info":{"description":"论衡 ThesisGuard — Personal Trading Research & Decision System","title":"ThesisGuard","version":"0.1.0"},"openapi":"3.1.0","paths":{"/":{"get":{"operationId":"root__get","responses":{"200":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"title":"Response Root  Get","type":"object"}}},"description":"Successful Response"}},"summary":"Root"}},"/api/v1/health":{"get":{"description":"Basic health check endpoint.","operationId":"health_check_api_v1_health_get","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}},"description":"Successful Response"}},"summary":"Health Check","tags":["health"]}},"/api/v1/health/db":{"get":{"description":"Database health check.","operationId":"health_check_db_api_v1_health_db_get","responses":{"200":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"title":"Response Health Check Db Api V1 Health Db Get","type":"object"}}},"description":"Successful Response"}},"summary":"Health Check Db","tags":["health"]}},"/api/v1/instruments/search":{"get":{"description":"Search instruments by symbol, name, alias, or tag.","operationId":"search_instrument_endpoint_api_v1_instruments_search_get","parameters":[{"in":"query","name":"query","required":true,"schema":{"maxLength":128,"minLength":1,"title":"Query","type":"string"}},{"in":"query","name":"limit","required":false,"schema":{"default":10,"maximum":50,"minimum":1,"title":"Limit","type":"integer"}}],"responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/InstrumentRead"},"title":"Response Search Instrument Endpoint Api V1 Instruments Search Get","type":"array"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Search Instrument Endpoint","tags":["instrument"]}},"/api/v1/instruments/{symbol}":{"get":{"description":"Get an instrument by symbol.","operationId":"get_instrument_endpoint_api_v1_instruments__symbol__get","parameters":[{"in":"path","name":"symbol","required":true,"schema":{"title":"Symbol","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InstrumentRead"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Get Instrument Endpoint","tags":["instrument"]}},"/api/v1/research/instruments/{instrument_id}/packages":{"get":{"description":"List all Research Package versions for an instrument.","operationId":"list_research_package_history_endpoint_api_v1_research_instruments__instrument_id__packages_get","parameters":[{"in":"path","name":"instrument_id","required":true,"schema":{"title":"Instrument Id","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ResearchPackageRead"},"title":"Response List Research Package History Endpoint Api V1 Research Instruments  Instrument Id  Packages Get","type":"array"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"List Research Package History Endpoint","tags":["research"]},"post":{"description":"Create initial Research Package version 1 for an instrument.","operationId":"create_initial_research_package_endpoint_api_v1_research_instruments__instrument_id__packages_post","parameters":[{"in":"path","name":"instrument_id","required":true,"schema":{"title":"Instrument Id","type":"string"}},{"description":"Required idempotency key for Research Package mutations.","in":"header","name":"Idempotency-Key","required":true,"schema":{"description":"Required idempotency key for Research Package mutations.","maxLength":128,"minLength":1,"title":"Idempotency-Key","type":"string"}}],"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResearchPackageRead"}}},"description":"Successful Response"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResearchErrorResponse"}}},"description":"Research resource not found"},"409":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResearchErrorResponse"}}},"description":"Research write, version, or idempotency conflict"},"422":{"content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/schemas/HTTPValidationError"},{"$ref":"#/components/schemas/ResearchErrorResponse"}]}}},"description":"FastAPI request validation or Research domain validation error"}},"summary":"Create Initial Research Package Endpoint","tags":["research"]}},"/api/v1/research/instruments/{instrument_id}/packages/current":{"get":{"description":"Get the current Research Package version for an instrument.","operationId":"get_current_research_package_endpoint_api_v1_research_instruments__instrument_id__packages_current_get","parameters":[{"in":"path","name":"instrument_id","required":true,"schema":{"title":"Instrument Id","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResearchPackageRead"}}},"description":"Successful Response"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResearchErrorResponse"}}},"description":"Research resource not found"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Get Current Research Package Endpoint","tags":["research"]}},"/api/v1/research/instruments/{instrument_id}/packages/refresh":{"post":{"description":"Create an append-only incremental Research Package version.","operationId":"refresh_research_package_endpoint_api_v1_research_instruments__instrument_id__packages_refresh_post","parameters":[{"in":"path","name":"instrument_id","required":true,"schema":{"title":"Instrument Id","type":"string"}},{"description":"Required idempotency key for Research Package mutations.","in":"header","name":"Idempotency-Key","required":true,"schema":{"description":"Required idempotency key for Research Package mutations.","maxLength":128,"minLength":1,"title":"Idempotency-Key","type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResearchRefreshRequest"}}},"required":true},"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResearchPackageRead"}}},"description":"Successful Response"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResearchErrorResponse"}}},"description":"Research resource not found"},"409":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResearchErrorResponse"}}},"description":"Research write, version, or idempotency conflict"},"422":{"content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/schemas/HTTPValidationError"},{"$ref":"#/components/schemas/ResearchErrorResponse"}]}}},"description":"FastAPI request validation or Research domain validation error"}},"summary":"Refresh Research Package Endpoint","tags":["research"]}},"/api/v1/research/instruments/{instrument_id}/packages/versions/{version}":{"get":{"description":"Get a specified Research Package version for an instrument.","operationId":"get_research_package_version_endpoint_api_v1_research_instruments__instrument_id__packages_versions__version__get","parameters":[{"in":"path","name":"instrument_id","required":true,"schema":{"title":"Instrument Id","type":"string"}},{"in":"path","name":"version","required":true,"schema":{"minimum":1,"title":"Version","type":"integer"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResearchPackageRead"}}},"description":"Successful Response"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResearchErrorResponse"}}},"description":"Research resource not found"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"FastAPI request validation error"}},"summary":"Get Research Package Version Endpoint","tags":["research"]}},"/api/v1/system/status":{"get":{"description":"Comprehensive system status check.\n\nReturns status of all core services: postgres, redis, object_storage, worker.","operationId":"system_status_api_v1_system_status_get","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SystemStatusResponse"}}},"description":"Successful Response"}},"summary":"System Status","tags":["system"]}},"/api/v1/system/status/postgres":{"get":{"description":"PostgreSQL specific status.","operationId":"postgres_status_api_v1_system_status_postgres_get","responses":{"200":{"content":{"application/json":{"schema":{"additionalProperties":true,"title":"Response Postgres Status Api V1 System Status Postgres Get","type":"object"}}},"description":"Successful Response"}},"summary":"Postgres Status","tags":["system"]}},"/api/v1/system/status/redis":{"get":{"description":"Redis specific status.","operationId":"redis_status_api_v1_system_status_redis_get","responses":{"200":{"content":{"application/json":{"schema":{"additionalProperties":true,"title":"Response Redis Status Api V1 System Status Redis Get","type":"object"}}},"description":"Successful Response"}},"summary":"Redis Status","tags":["system"]}},"/api/v1/system/status/storage":{"get":{"description":"MinIO/Object Storage specific status.","operationId":"storage_status_api_v1_system_status_storage_get","responses":{"200":{"content":{"application/json":{"schema":{"additionalProperties":true,"title":"Response Storage Status Api V1 System Status Storage Get","type":"object"}}},"description":"Successful Response"}},"summary":"Storage Status","tags":["system"]}},"/api/v1/tasks/echo":{"post":{"description":"Dispatch an echo task to the worker (for testing).","operationId":"dispatch_echo_api_v1_tasks_echo_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EchoRequest"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskDispatchResponse"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Dispatch Echo","tags":["tasks"]}},"/api/v1/tasks/health-check":{"post":{"description":"Dispatch a system health check task to the worker.","operationId":"dispatch_health_check_api_v1_tasks_health_check_post","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskDispatchResponse"}}},"description":"Successful Response"}},"summary":"Dispatch Health Check","tags":["tasks"]}},"/api/v1/watchlist":{"get":{"description":"List watchlist items.","operationId":"list_watchlist_endpoint_api_v1_watchlist_get","parameters":[{"in":"query","name":"classification","required":false,"schema":{"anyOf":[{"maxLength":32,"type":"string"},{"type":"null"}],"title":"Classification"}}],"responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/WatchlistItemRead"},"title":"Response List Watchlist Endpoint Api V1 Watchlist Get","type":"array"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"List Watchlist Endpoint","tags":["watchlist"]},"post":{"description":"Add an instrument to watchlist and classify it.","operationId":"add_watchlist_endpoint_api_v1_watchlist_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WatchlistAddRequest"}}},"required":true},"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WatchlistItemRead"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Add Watchlist Endpoint","tags":["watchlist"]}},"/api/v1/watchlist/{item_id}":{"delete":{"description":"Delete a watchlist item.","operationId":"delete_watchlist_endpoint_api_v1_watchlist__item_id__delete","parameters":[{"in":"path","name":"item_id","required":true,"schema":{"title":"Item Id","type":"string"}}],"responses":{"204":{"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Delete Watchlist Endpoint","tags":["watchlist"]},"patch":{"description":"Update a watchlist item.","operationId":"update_watchlist_endpoint_api_v1_watchlist__item_id__patch","parameters":[{"in":"path","name":"item_id","required":true,"schema":{"title":"Item Id","type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WatchlistUpdateRequest"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WatchlistItemRead"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Update Watchlist Endpoint","tags":["watchlist"]}}}}